math floor and ceil round to int rather than float #8258

+++ crates/nu-command/src/math/floor.rs
@@ -58,8 +58,8 @@ impl Command for SubCommand {
 fn operate(value: Value, head: Span) -> Value {
     match value {
         Value::Int { .. } => value,
-        Value::Float { val, span } => Value::Float {
-            val: val.floor(),
+        Value::Float { val, span } => Value::Int {
+            val: val.floor() as i64,
             span,
         },
         Value::Error { .. } => value,
